commit ne

+++ DATN-TOPDEV-2024/BE/backEndSpringBoot/src/main/java/com/be/controller/BrandController.java
@@ -1,9 +1,9 @@
 package com.be.controller;
 
-
 import com.be.entity.Brand;
 import com.be.service.BrandService;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,31 +12,56 @@ import java.util.List;
 @RequestMapping("/api/brands")
 public class BrandController {
 
-    @Autowired
-    private BrandService brandService;
+    private final BrandService brandService;
 
+    public BrandController(BrandService brandService) {
+        this.brandService = brandService;
+    }
+
+    // Lấy danh sách tất cả các Brand
     @GetMapping
-    public List<Brand> getAllBrands() {
-        return brandService.getAllBrands();
+    public ResponseEntity<List<Brand>> getAllBrands() {
+        List<Brand> brands = brandService.getAllBrands();
+        return ResponseEntity.ok(brands);
     }
 
+    // Lấy thông tin Brand theo ID
     @GetMapping("/{id}")
-    public Brand getBrandById(@PathVariable Long id) {
-        return brandService.getBrandById(id);
+    public ResponseEntity<Brand> getBrandById(@PathVariable Long id) {
+        Brand brand = brandService.getBrandById(id);
+        if (brand != null) {
+            return ResponseEntity.ok(brand);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
+    // Tạo mới một Brand
     @PostMapping
-    public Brand createBrand(@RequestBody Brand brand) {
-        return brandService.createBrand(brand);
+    public ResponseEntity<Brand> createBrand(@RequestBody Brand brand) {
+        Brand createdBrand = brandService.createBrand(brand);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdBrand);
     }
 
+    // Cập nhật thông tin một Brand theo ID
     @PutMapping("/{id}")
-    public Brand updateBrand(@PathVariable Long id, @RequestBody Brand brand) {
-        return brandService.updateBrand(id, brand);
+    public ResponseEntity<Brand> updateBrand(@PathVariable Long id, @RequestBody Brand brand) {
+        Brand updatedBrand = brandService.updateBrand(id, brand);
+        if (updatedBrand != null) {
+            return ResponseEntity.ok(updatedBrand);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
+    // Xóa một Brand theo ID
     @DeleteMapping("/{id}")
-    public void deleteBrand(@PathVariable Long id) {
-        brandService.deleteBrand(id);
+    public ResponseEntity<Void> deleteBrand(@PathVariable Long id) {
+        if (brandService.getBrandById(id) != null) {
+            brandService.deleteBrand(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 }
